requirements.txt
torch
torchvision
timm
numpy<2
pillow_heif
matplotlib
flask
tqdm
requests 